# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Auron Business OS API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /events:
    get:
      operationId: listEvents
      tags: [events]
      summary: List events
      parameters:
        - name: search
          in: query
          required: false
          schema:
            type: string
        - name: status
          in: query
          required: false
          schema:
            type: string
        - name: eventType
          in: query
          required: false
          schema:
            type: string
        - name: clientId
          in: query
          required: false
          schema:
            type: number
        - name: fromDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: toDate
          in: query
          required: false
          schema:
            type: string
            format: date
        - name: page
          in: query
          required: false
          schema:
            type: number
            default: 1
        - name: limit
          in: query
          required: false
          schema:
            type: number
            default: 50
      responses:
        "200":
          description: Events list
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventListResponse"
    post:
      operationId: createEvent
      tags: [events]
      summary: Create a new event
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
  /events/{id}:
    get:
      operationId: getEvent
      tags: [events]
      summary: Get an event by ID with full profitability
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Event with profitability
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventDetail"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorEnvelope"
    patch:
      operationId: updateEvent
      tags: [events]
      summary: Update an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventUpdate"
      responses:
        "200":
          description: Updated event
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Event"
    delete:
      operationId: deleteEvent
      tags: [events]
      summary: Delete an event
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # Event Revenue
  /events/{eventId}/revenue:
    get:
      operationId: getEventRevenue
      tags: [eventRevenue]
      summary: Get revenue record for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Event revenue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRevenue"
    post:
      operationId: upsertEventRevenue
      tags: [eventRevenue]
      summary: Create or update event revenue
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventRevenueInput"
      responses:
        "200":
          description: Updated revenue
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventRevenue"

  # Event Costs
  /events/{eventId}/costs:
    get:
      operationId: listEventCosts
      tags: [eventCosts]
      summary: List all costs for an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
      responses:
        "200":
          description: Event costs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/EventCost"
    post:
      operationId: createEventCost
      tags: [eventCosts]
      summary: Add a cost to an event
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCostInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventCost"
  /events/{eventId}/costs/{id}:
    patch:
      operationId: updateEventCost
      tags: [eventCosts]
      summary: Update a cost entry
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EventCostUpdate"
      responses:
        "200":
          description: Updated cost
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EventCost"
    delete:
      operationId: deleteEventCost
      tags: [eventCosts]
      summary: Delete a cost entry
      parameters:
        - name: eventId
          in: path
          required: true
          schema:
            type: number
        - name: id
          in: path
          required: true
          schema:
            type: number
      responses:
        "204":
          description: Deleted

  # Leads
...
components:
...
  schemas:
...
    ErrorEnvelope:
      type: object
      required: [error]
      properties:
        error:
          type: string

    # Dashboard schemas
...
    Event:
      type: object
      required: [id, name, clientId, eventType, status, eventDate, createdAt]
      properties:
        id:
          type: number
        name:
          type: string
        clientId:
          type: number
        clientName:
          type: ["string", "null"]
        eventType:
          type: string
        status:
          type: string
          enum: [upcoming, in_progress, completed, cancelled]
        eventDate:
          type: string
          format: date
        venue:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        salespersonId:
          type: ["integer", "null"]
        salespersonName:
          type: ["string", "null"]
        operationsManagerId:
          type: ["integer", "null"]
        leadSource:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isDemo:
          type: boolean
        totalRevenue:
          type: number
        totalCost:
          type: number
        grossProfit:
          type: number
        grossMarginPct:
          type: number
        profitabilityIndicator:
          type: string
          enum: [excellent, healthy, warning, loss, awaiting_data]
        totalCollected:
          type: number
        totalOutstanding:
          type: number
        createdAt:
          type: string
          format: date-time
    EventDetail:
      type: object
      required: [id, name, clientId, eventType, status, eventDate, createdAt, revenue, costs]
      properties:
        id:
          type: number
        name:
          type: string
        clientId:
          type: number
        clientName:
          type: ["string", "null"]
        eventType:
          type: string
        status:
          type: string
          enum: [upcoming, in_progress, completed, cancelled]
        eventDate:
          type: string
          format: date
        venue:
          type: ["string", "null"]
        location:
          type: ["string", "null"]
        salespersonId:
          type: ["integer", "null"]
        salespersonName:
          type: ["string", "null"]
        operationsManagerId:
          type: ["integer", "null"]
        leadSource:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        isDemo:
          type: boolean
        totalRevenue:
          type: number
        totalCost:
          type: number
        grossProfit:
          type: number
        grossMarginPct:
          type: number
        profitabilityIndicator:
          type: string
        totalCollected:
          type: number
        totalOutstanding:
          type: number
        createdAt:
          type: string
          format: date-time
        revenue:
          $ref: "#/components/schemas/EventRevenue"
        costs:
          type: array
          items:
            $ref: "#/components/schemas/EventCost"
    EventInput:
      type: object
      required: [name, clientId, eventType, status, eventDate]
      properties:
        name:
          type: string
          minLength: 1
        clientId:
          type: number
        eventType:
          type: string
        status:
          type: string
          enum: [upcoming, in_progress, completed, cancelled]
        eventDate:
          type: string
          format: date
        venue:
          type: string
        location:
          type: string
        salespersonId:
          type: number
        operationsManagerId:
          type: number
        leadSource:
          type: string
        notes:
          type: string
    EventUpdate:
      type: object
      properties:
        name:
          type: string
          minLength: 1
        clientId:
          type: number
        eventType:
          type: string
        status:
          type: string
          enum: [upcoming, in_progress, completed, cancelled]
        eventDate:
          type: string
          format: date
        venue:
          type: string
        location:
          type: string
        salespersonId:
          type: number
        operationsManagerId:
          type: number
        leadSource:
          type: string
        notes:
          type: string
    EventListResponse:
      type: object
      required: [data, total, page, limit]
      properties:
        data:
          type: array
          items:
            $ref: "#/components/schemas/Event"
        total:
          type: number
        page:
          type: number
        limit:
          type: number

    # Event Revenue schemas
    EventRevenue:
      type: object
      required: [id, eventId, contractValue, gst, totalInvoiceValue, netRevenue, advanceReceived, secondPayment, finalPayment, totalCollected, outstandingAmount, createdAt]
      properties:
        id:
          type: number
        eventId:
          type: number
        contractValue:
          type: number
        discount:
          type: number
        gst:
          type: number
        totalInvoiceValue:
          type: number
        netRevenue:
          type: number
        advanceReceived:
          type: number
        secondPayment:
          type: number
        finalPayment:
          type: number
        totalCollected:
          type: number
        outstandingAmount:
          type: number
        paymentStatus:
          type: string
          enum: [pending, partially_paid, paid, overdue]
        invoiceNumber:
          type: ["string", "null"]
        dueDate:
          type: ["string", "null"]
          format: date
        createdAt:
          type: string
          format: date-time
    EventRevenueInput:
      type: object
      required: [contractValue, gst]
      properties:
        contractValue:
          type: number
        discount:
          type: number
        gst:
          type: number
        advanceReceived:
          type: number
        secondPayment:
          type: number
        finalPayment:
          type: number
        paymentStatus:
          type: string
          enum: [pending, partially_paid, paid, overdue]
        invoiceNumber:
          type: string
        dueDate:
          type: string
          format: date

    # Event Cost schemas
    EventCost:
      type: object
      required: [id, eventId, category, amount, gst, paymentStatus, createdAt]
      properties:
        id:
          type: number
        eventId:
          type: number
        vendorId:
          type: ["integer", "null"]
        vendorName:
          type: ["string", "null"]
        category:
          type: string
        description:
          type: ["string", "null"]
        amount:
          type: number
        gst:
          type: number
        totalAmount:
          type: number
        paymentStatus:
          type: string
          enum: [pending, paid, partially_paid]
        date:
          type: ["string", "null"]
          format: date
        referenceNumber:
          type: ["string", "null"]
        createdAt:
          type: string
          format: date-time
    EventCostInput:
      type: object
      required: [category, amount]
      properties:
        vendorId:
          type: number
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        gst:
          type: number
        paymentStatus:
          type: string
          enum: [pending, paid, partially_paid]
        date:
          type: string
          format: date
        referenceNumber:
          type: string
    EventCostUpdate:
      type: object
      properties:
        vendorId:
          type: number
        category:
          type: string
        description:
          type: string
        amount:
          type: number
        gst:
          type: number
        paymentStatus:
          type: string
          enum: [pending, paid, partially_paid]
        date:
          type: string
          format: date
        referenceNumber:
          type: string

    # Lead schemas
...